All downhill from here.
This repo initially started as a fork of [ZaneyOS](https://gitlab.com/Zaney/zaneyos) and,
while that project has a great-looking Hyprland setup, I wanted to make structural
changes that warranted a fork. In the time since, the config has diverged enough that it
makes sense to squash all the commits that got it to this base level and start making more
atomic commits.